% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_plot.R
\name{agg_plot}
\alias{agg_plot}
\title{Plot aggregated fx swap data}
\usage{
agg_plot(
  df,
  series = "netto",
  counterparty = "all",
  reporting_banks = "all",
  tenors = "all",
  currency = levels(as.factor(df$FxFrgnCcy)),
  grouping = "tenor",
  start_date = "2020-03-01",
  end_date = Sys.Date() - 1,
  yrange = NULL
)
}
\arguments{
\item{df}{Data frame med data som skal plottes}

\item{series}{hvilken sluttserie skal vises? Tar verdiene "netto", "innlan" og "utlan"}

\item{tenors}{Hvilke løpetider skal vises? Default er alle i datasettet som brukes som input}

\item{grouping}{Hvilken variabel skal grupperes over? Default er løpetid ("tenor")}

\item{yrange}{egendefinert y-range (optional)}

\item{counterparties}{Velg hvilke motparter som skal plottes mot}

\item{currencies}{Hvilke valutaer skal inkluderes? Default er alle}

\item{date_start}{startdato for graf}

\item{date_end}{sluttdato for graf}
}
\value{

}
\description{
Plot aggregated fx swap data
}
\examples{
\dontrun{
agglist <- noma_agg_plot_fx(fx, "FOREIGN", "netto", c("3m", "6m"))
agglist$p
}
}
