% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deposits.R
\name{deposits}
\alias{deposits}
\title{Hent data for usikrede transaksjoner}
\usage{
deposits(
  start_date = "2019-11-30",
  counterparties = "all",
  netting = FALSE,
  ctrInfo = TRUE,
  keepON = TRUE,
  additional_fields = c()
)
}
\arguments{
\item{start_date}{character. Start-dato for uttrekk. Angis som "åååå-mm-dd"}

\item{counterparties}{vector. tar tre verdier: \cr
\itemize{
\item "all" (default) gir all rapportert omsetning mot alle motparter
\item reportingagents" gir omsetning mellom rapportørbanker
\item "nonreportingagents" gir omsetning mellom rapportørbanker og ikke-rapportører
}}

\item{netting}{logical. Hvis TRUE (default), inkluderes kun lending- \cr
transaksjoner mellom rapportørbanker.}

\item{ctrInfo}{logical. Hvis TRUE(default), inkluder navn og leikode for motparter}

\item{keepON}{logical. Hvis TRUE, inkluder ON-transaksjnoer}

\item{additional_fields}{char. En vektor med kolonnenavn fra NBIP-databasen \cr
man vil trekke ut i tillegg til standardkolonnene.}
}
\description{
Hent data for usikrede transaksjoner
}
\examples{

## Hent ut med standardvalg:
depos <- deposits()

}
