% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_swap.R
\encoding{UTF-8}
\name{fx_swap}
\alias{fx_swap}
\title{Hent FX-swap-data fra RPD}
\usage{
fx_swap(
  start_date = "2019-11-30",
  counterparties = "all",
  netting = FALSE,
  ctrInfo = TRUE,
  additional_fields = c()
)
}
\arguments{
\item{start_date}{character. Start-dato for uttrekk. Angis som "åååå-mm-dd"}

\item{counterparties}{vector. tar tre verdier: \cr
\itemize{
\item "all" (default) gir all rapportert omsetning mot alle motparter
\item reportingagents" gir omsetning mellom rapportørbanker
\item "nonreportingagents" gir omsetning mellom rapportørbanker og ikke-rapportører
}}

\item{netting}{logical. Hvis TRUE, kun transaksjoner markert "SELL" (dvs. utlån av kroner) mellom \cr
rapportørbanker blir hentet ut. Unngår da dobbelttelling av inn- og utlån mellom  rapportørbanker \cr
FALSE er default}

\item{ctrInfo}{logical. Hvis TRUE(default), inkluder navn og leikode for motparter}

\item{additional_fields}{char. En vektor med kolonnenavn fra NBIP-databasen man vil \cr
trekke ut i tillegg til standardkolonnene.}
}
\description{
\code{fx_swap} henter data fra NBIP-databasen (der data fra RPD lagres).
Valgte parametre styrer hvilke data som inkluderes i uttrekket
}
\examples{
## Hent ut data med default-verdier. (Normalt det man har mest bruk for)
fx <- fx_swap()
##  Hent ut kun omsetning mellom bankene som rapporterer, nett transaksjoner mellom dem,
## og sett startdato til 30. juni 2020.
fx <- fx_swap(start_date = "2020-06-30",
counterparties = "reportingagents", netting = TRUE)
}
