% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_holidays.R
\encoding{UTF-8}
\name{import_holidays}
\alias{import_holidays}
\title{Importer liste over norske helligdager}
\usage{
import_holidays()
}
\description{
\code{import_holidays} henter en liste over norske helligdager fra NBDataHub.
}
\examples{
helligdager <- import_holidays()
}
