% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregate_fx_swap_outstanding.R
\encoding{UTF-8}
\name{plot_aggregate_fx_swap_outstanding}
\alias{plot_aggregate_fx_swap_outstanding}
\title{Plot aggregert utestående i FX-swap}
\usage{
plot_aggregate_fx_swap_outstanding(
  df,
  start_date = "2020-03-01",
  end_date = Sys.Date(),
  reporting_banks = "all",
  counterparty = "all_other_cps",
  tenors = "all",
  show_net = TRUE,
  currency = "USD",
  return_data = FALSE,
  title = TRUE,
  subtitle = TRUE,
  title_text = NULL,
  subtitle_text = NULL
)
}
\arguments{
\item{df}{dataframe. Datasett med fx-swapdata.}

\item{start_date}{char. Start-dato for grafer. På format "åååå-mm-dd"}

\item{end_date}{char. Slutt-dato for grafer. På format "åååå-mm-dd"}

\item{reporting_banks}{char. Hvilke rapportørbanker man skal plotte utestående volum for
tar verdien "all" for alle banker, eller en vektor ett eller flere banknavn}

\item{counterparty}{char. Hvilke motparter man skal plotte utestående volum mot
Tar et motpartsnavn eller en av tre verdier:
\itemize{
\item "all_other_cps": Alle motparter bortsett fra andre rapportørbanker (default)
\item "foreign_cps": Utenlandske motparter
\item "no_cps": Norske motparter (utenom rapportørbanker)
}
\cr
Merk at om man velger et motpartsnavn er ikke automatisk tittel støttet. Den må da
spesifiseres selv med title_text.}

\item{tenors}{char. Hvilke løpetider man vil inkludere i utestående.
Tar enten verdien "all" (default) eller en vektor med  en eller flere valgte løpetider}

\item{show_net}{logical. Skal man vise linje for netto utestående? Tar verdier TRUE eller FALSE}

\item{currency}{char. Hvilken valuta man skal vise grafer for?
Tar valutakode spesifisert med blokkbokstaver. "USD" er default}

\item{return_data}{logical. Skal man få ut de underliggende dataene? Tar verdiene FALSE (defualt)
og TRUE. Velger man TRUE, blir output en liste som inneholder graf og underliggende data}

\item{title}{logical. Skal man lage tittel på grafen? Tar verdier TRUE (default) eller FALSE}

\item{subtitle}{logical. Skal man lage undertittel på grafen? Tar verdier TRUE (default) eller FALSE}

\item{title_text}{NULL eller char. Egendefinert tittel på graf. Hvis ikke spesifisert, lages en automatisk.}

\item{subtitle_text}{NULL eller char. Egendefinert undertittel på graf. Hvis ikke spesifisert,
lages en automatisk.}
}
\description{
Plot aggregert utestående i FX-swap
}
\examples{
\dontrun{
Med default-verdier og datasett med fx-swap-data som heter "fx":
p = plot_aggregate_fx_swap_outstanding(df = fx)
Kun for Danske og DNB, utenlandske motparter, 1 og 3 måneders løpetid:
fx \%>\% plot_aggregate_fx_swap_outstanding(df = fx,
reporting_banks = c("Danske Bank", "DNB Bank ASA"),
tenors = c("1m", "3m"))

}


}
