% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fx_swap_outstanding.R
\name{plot_fx_swap_outstanding}
\alias{plot_fx_swap_outstanding}
\title{Plot Utestående i FX-swap}
\usage{
plot_fx_swap_outstanding(
  df,
  start_date = "2020-03-01",
  end_date = Sys.Date(),
  reporting_banks = c("DNB Bank ASA"),
  grouping = "tenor",
  grouping_restriction = FALSE,
  truncate_names = TRUE,
  return_data = FALSE,
  counterparty = "all",
  currency = c("USD"),
  tenors = "all",
  title = TRUE,
  subtitle = TRUE,
  title_text = NULL,
  subtitle_text = NULL
)
}
\arguments{
\item{df}{Dataframe. Datasett med fx-swap-data.}

\item{start_date}{character. Start-dato for grafer. På format "åååå-mm-dd"}

\item{end_date}{character. Slutt-dato for grafer. På format "åååå-mm-dd"}

\item{reporting_banks}{character vector. Hvilke rapportørbanker man skal plotte utestående volum for
tar verdien "all" for alle banker, eller en vektor ett eller flere
banknavn. "DNB Bank ASA" er default.}

\item{grouping}{character. Hvilken variabler skal man gruppere etter. Default er "tenor".
Man får da omsetning fordelt på løpetid. Tar i prinsippet hvilket som helst kolonnenavn i datasettet.
Generering av automatiske titler er støttet for "tenor", "CounterPartyName", "CounterPartyCountry", "ReportingAgentName",
og "FxFrgnCcy".}

\item{grouping_restriction}{FALSE eller integer. Skal man begrense antall grupperingsvariabler
man skal vise data for? Tar verdien FALSE eller et tall. Velger man f.eks.
grouping = "CounterPartyName", og grouping_restriction = 5, plottes omsetning
for de fem største motpartene på inn- og utlånssiden. Omsetning for øvrige
motparter klassifiseres som "Andre".}

\item{truncate_names}{locigal. Skal man trunkere navn? Ofte nyttig, fordi mange motparter har lange navn.}

\item{return_data}{logical. Skal man få ut de underliggende dataene? Tar verdiene FALSE (defualt)
og TRUE. Velger man TRUE, blir output en liste som inneholder graf og underliggende data}

\item{counterparty}{character. Hvilke motparter man skal plotte utestående volum mot?
Tar en vektor med motpartsnavn eller en av tre verdier:
\itemize{
\item "all_other_cps": Alle motparter bortsett fra andre rapportørbanker (default)
\item "foreign_cps": Utenlandske motparter
\item "no_cps": Norske motparter (utenom rapportørbanker)
}}

\item{currency}{character. Hvilken valuta man skal vise grafer for. Tar valutakode spesifisert
med blokkbokstaver. "USD" er default
#' @param return_data logical. Skal man få ut de underliggende dataene? Tar verdiene FALSE (defualt)
og TRUE. Velger man TRUE, blir output en liste som inneholder graf og underliggende data}

\item{tenors}{character vector. hvilke løpetider man vil inkludere i utestående. Tar enten verdien
"all" (default) eller en vektor med  en eller flere valgte løpetider}

\item{title}{logical. Skal man lage tittel på grafen? Tar verdier TRUE (default) eller FALSE}

\item{subtitle}{logical. Skal man lage undertittel på grafen? Tar verdier TRUE (default) eller FALSE}

\item{title_text}{NULL eller char. Egendefinert tittel på graf. Hvis ikke spesifisert, lages en automatisk.}

\item{subtitle_text}{NULL eller char. Egendefinert undertittel på graf. Hvis ikke spesifisert,
lages en automatisk.}
}
\description{
Plot Utestående i FX-swap
}
\examples{
\dontrun{
#Utestående for DNB fordelt på løpetid
p <- plot_fx_swap_outstanding(fx, reporting_banks = "DNB Bank ASA")
p
#Utestående for DNB fordelt på 5 største motparter på inn- og utlånssiden
p <- plot_fx_swap_outstanding(fx, reporting_banks = "DNB Bank ASA",
                              grouping  = "CounterPartyName",
                              grouping_restriction = 5)
p
#Utestående for DNB fordelt på 5 største motparter på inn- og
#utlånssiden for kun 3m-swapper:
p <- plot_fx_swap_outstanding(fx, reporting_banks = "DNB Bank ASA",
                              tenors = "3m",
                              grouping  = "CounterPartyName",
                              grouping_restriction = 5)
p

}
}
