% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intraday_fx_swap_trades.R
\encoding{UTF-8}
\name{plot_intraday_fx_swap_trades}
\alias{plot_intraday_fx_swap_trades}
\title{Plot intradag-omsetning i FX-swapper}
\usage{
plot_intraday_fx_swap_trades(
  df,
  start_date,
  end_date = Sys.Date(),
  time_start = NULL,
  reporting_banks = "DNB Bank ASA",
  tenors = "TN",
  currency = "USD",
  counterparty = "all",
  grouping = "tenor",
  truncate_names = FALSE,
  return_data = FALSE,
  title = TRUE,
  subtitle = TRUE,
  title_text = NULL,
  subtitle_text = NULL
)
}
\arguments{
\item{df}{Datasett med FX-swap-data}

\item{start_date}{char. Start-dato for plot.}

\item{end_date}{char. Slutt-dato for plot.}

\item{time_start}{char. Fra hvilket tidspunkt vil man vise omsetning. Kan settes til f.eks.
"07:00" for å kun vise omsetning etter 07:00.}

\item{reporting_banks}{char. character vector. Hvilke rapportørbanker man skal plotte utestående volum for \cr
tar verdien "all" for alle banker, eller en vektor ett eller flere \cr
banknavn. "DNB Bank ASA" er default.}

\item{tenors}{character vector. hvilke løpetider man vil inkludere i utestående. Tar enten verdien \cr
"all" (default) eller en vektor med  en eller flere valgte løpetider}

\item{currency}{character. Hvilken valuta man skal vise grafer for. Tar valutakode spesifisert \cr
med blokkbokstaver. "USD" er default}

\item{counterparty}{character. Hvilke motparter man skal plotte utestående volum mot. \cr
Tar verdien "all" (default) for alle motparter, eller en vektor med  navn på motparter.}

\item{grouping}{character. Hvilken variabel skal man gruppere etter. Default er "tenor".
Man får da omsetning fordelt på løpetid. \cr Tar også verdiene CounterPartyName \cr
og CounterpartyCountry. Det gir omsetning fordelt på hhv. motpart og motpartsland}

\item{truncate_names}{locigal. Skal man trunkere navn? Ofte nyttig, fordi mange motparter har lange navn.}

\item{return_data}{logical. Skal man få ut de underliggende dataene? Tar verdiene FALSE (defualt) \cr
og TRUE. Velger man TRUE, blir output en liste som inneholder graf og underliggende data}

\item{title}{logical. Skal man lage tittel på grafen? Tar verdier TRUE (default) eller FALSE}

\item{subtitle}{logical. Skal man lage undertittel på grafen? Tar verdier TRUE (default) eller FALSE}

\item{title_text}{NULL eller char. Egendefinert tittel på graf. Hvis ikke spesifisert, lages en automatisk.}

\item{subtitle_text}{NULL eller char. Egendefinert undertittel på graf. Hvis ikke spesifisert, \cr
lages en automatisk.}
}
\description{
Plot intradag-omsetning i FX-swapper
}
\examples{
\dontrun{

#Plot omsetning etter løpetid:
plot_intraday_fx_swap_trades(fx,
                             start_date = "2020-11-11",
                             tenors = "all",
                             reporting_banks = "all",
                             grouping = "tenor")

#Plot omsetning etter rapporterende bank:
plot_intraday_fx_swap_trades(fx,
                             start_date = "2020-11-11",
                             tenors = "all",
                             reporting_banks = "all",
                             grouping = "ReportingAgentName")

#Plot omsetning etter rapporterende bank, og kun for 3m-swapper:
plot_intraday_fx_swap_trades(fx,
                             start_date = "2020-11-11",
                             tenors = "3m",
                             reporting_banks = "all",
                             grouping = "ReportingAgentName")

}

}
