% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo}
\alias{repo}
\title{Hent inn repo-data fra RPD}
\usage{
repo(
  start_date = "2019-11-30",
  counterparties = "all",
  netting = FALSE,
  ctrInfo = TRUE,
  additional_fields = c()
)
}
\arguments{
\item{start_date}{character. Start-dato for uttrekk. Angis som "åååå-mm-dd"}

\item{counterparties}{vector. tar tre verdier: \cr
\itemize{
\item "all" (default) gir all rapportert omsetning mot alle motparter
\item reportingagents" gir omsetning mellom rapportørbanker
\item "nonreportingagents" gir omsetning mellom rapportørbanker og ikke-rapportører
}}

\item{netting}{logical. Hvis TRUE, kun transaksjoner markert "SELL" (dvs. utlån av kroner) mellom \cr
rapportørbanker blir hentet ut. Unngår da dobbelttelling av inn- og utlån mellom  rapportørbanker \cr
FALSE er default}

\item{ctrInfo}{logical. Hvis TRUE(default), inkluder navn og leikode for motparter}

\item{additional_fields}{vector. En vektor med kolonnenavn fra NBIP-databasen man vil \cr
trekke ut i tillegg til standardkolonnene.}
}
\description{
Hent inn repo-data fra RPD
}
\examples{
## Hent ut data med default-verdier. (Normalt det man har mest bruk for)
repos <- repo()
## Hent ut kun omsetning mellom bankene som rapporterer, siden 30. juni 2020,
## og nett transaksjoner mellom dem:
repos <- repo(start_date = "2020-06-30",
counterparties = "reportingagents", netting = TRUE)
}
