% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpd_dashboard.R
\name{rpd_dashboard}
\alias{rpd_dashboard}
\title{Dashboard for RPD-data}
\usage{
rpd_dashboard(load_source_data = TRUE)
}
\arguments{
\item{load_source_data}{logical. Skal man laste inn data til applikasjonen eller ikke.
Dette kan ta litt tid, avhengig av hastighet på tilkobling etc.
Alternativ kan man laste ned datasett på forhånd, og kjøre applikasjon
med 'load_source_data = F'. Forhåndslastede datasett må da ha navnene
fx, repos og depos for hhv fx-swap, repo og usikrede transaksjoner."}
}
\description{
Dashboard for RPD-data
}
\examples{

#Start dashboard og last inn data
\dontrun{
rpd_dashboard()
}



}
