% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_names.R
\name{search_names}
\alias{search_names}
\title{Søk etter navn}
\usage{
search_names(v, name)
}
\arguments{
\item{v}{character vector. F.eks. en kolonne i en dataframe}

\item{name}{char. F.eks. "chase". Funksjonen er ikke case sensitiv, \cr
så det bare å søke med små eller store bokstaver som man vil}
}
\description{
Søk etter navn
}
\examples{
\dontrun{
#Søk etter navn i kolonnen for motparter i et FX-swapdatasett med navn "fx":
search_names(v = fx$CounterPartyName, name = "chase")
}
}
